+++ src/vue/modules/user-movements-history/store/index.js
@@ -21,7 +21,6 @@ export const actions = {
     return api.getWithSignature('/v3/history', {
       page: {
         order: 'desc',
-        limit: 10,
       },
       filter: {
         account: filters.accountId,