+++ src/vue/modules/movements-history/store/index.js
@@ -29,7 +29,6 @@ export const actions = {
     return api.getWithSignature('/v3/history', {
       page: {
         order: 'desc',
-        limit: 10,
       },
       filter: {
         account: rootGetters[vuexTypes.accountId],
@@ -48,7 +47,6 @@ export const actions = {
     return api.getWithSignature('/v3/movements', {
       page: {
         order: 'desc',
-        limit: 10,
       },
       filter: {
         asset: assetCode,