+++ CHANGELOG.md
@@ -11,6 +11,10 @@ for further information about branching and tagging conventions.
 - Localization mistakes
 - A bug with trailing digits count on amount fields
 
+### "Under the hood" changes
+#### Changed
+- Now using @tokend/js-sdk@1.10.1-rc.0
+
 ## [1.0.3-rc.2] - 2019-08-29
 #### Removed
 - Login redirect